assignment-2-784478/data/client/requirements.txt
pyyaml
apscheduler
pandas
minio
paho-mqtt